Move JSON RPC-related items into a new module and export it.
